// src/app/DailyAnalysis/models/daily.model.ts
export class DailyModel {
  time?: string;
  previousClose: number;
  prevPrevGap: number;
  open: number;
  prevOpenGap: number;
  high: number;
  openHighGap: number;
  prevHighGap: number;
  low: number;
  prevLowGap: number;
  highLowGap: number;
  lowLowGap: number;
  openLowGap: number;
  lowCloseGap: number;
  close: number;
  openCloseGap: number;
  closeToCloseGap: number;
  highCloseGap: number;
  volume: number;
}